Remove legacy seed data scripts and prestart.sh file; introduce new seed data structure and seeding script for improved organization and functionality
