+++ backend/app/scripts/seed_data/seed_data.json
@@ -1,0 +1,61 @@
+{
+    "organization": [
+      {
+        "id": 1001,
+        "name": "Project Tech4dev",
+        "is_active": true
+      }
+    ],
+    "projects": [
+      {
+        "id": 1002,
+        "name": "Glific",
+        "description": "Two way communication platform",
+        "is_active": true,
+        "organization_id": 1001
+      },
+      {
+        "id": 1003,
+        "name": "Dalgo",
+        "description": "Data platform for the social sector",
+        "is_active": true,
+        "organization_id": 1001
+      }
+    ],
+    "users": [
+      {
+        "id": "4629f304-10c5-45f8-80f5-26ea1066a1cb",
+        "email": "superuser@example.com",
+        "password": "changethis",
+        "full_name": "SUPERUSER",
+        "is_active": true,
+        "is_superuser": true
+      },
+      {
+        "id": "375503ac-f9d1-465e-8a87-be6974799ce1",
+        "email": "org_admin@example.com",
+        "password": "admin123",
+        "full_name": "ADMIN",
+        "is_active": true,
+        "is_superuser": false
+      }
+    ],
+    "apikeys": [
+      {
+        "id": 1,
+        "organization_id": 1001,
+        "user_id": "4629f304-10c5-45f8-80f5-26ea1066a1cb",
+        "api_key": "ApiKey No3x47A5qoIGhm0kVKjQ77dhCqEdWRIQZlEPzzzh7i8",
+        "is_deleted": false,
+        "deleted_at": null
+      },
+      {
+        "id": 2,
+        "organization_id": 1001,
+        "user_id": "375503ac-f9d1-465e-8a87-be6974799ce1",
+        "api_key": "ApiKey Px8y47B6roJHin1lWLkR88eiDrFdXSJRZmFQazzai8j9",
+        "is_deleted": false,
+        "deleted_at": null
+      }
+    ]
+  }
