+++ backend/scripts/prestart.sh
@@ -1,7 +1,4 @@
-#! /usr/bin/env bash
-
-set -e
-set -x
+#!/bin/bash
 
 # Let the DB start
 python app/backend_pre_start.py
@@ -9,12 +6,8 @@ python app/backend_pre_start.py
 # Run migrations
 alembic upgrade head
 
-# Initialize services
-services=(
-    app/initial_data.py
-    app/initial_storage.py
-)
+# Seed the database
+python scripts/seed_data/seed_data.py  # This runs the seeding process
 
-for i in ${services[@]}; do
-    python $i
-done
+# Create initial data in DB
+python app/initial_data.py
